+++ libs/angular/src/lib/button/button.component.ts
@@ -4,7 +4,7 @@ import {
   HostBinding,
   Input
 } from '@angular/core'
-import { ButtonType, ButtonVariant } from '@sebgroup/extract';
+import { ButtonSize, ButtonVariant } from '@sebgroup/extract';
 
 @Component({
   // eslint-disable-next-line @angular-eslint/component-selector
@@ -15,9 +15,9 @@ import { ButtonType, ButtonVariant } from '@sebgroup/extract';
 export class NggButtonComponent {
 
   @Input() variant?: ButtonVariant;
-  @Input() size?: ButtonType;
+  @Input() size?: ButtonSize;
 
   @HostBinding('class') get classes(): string {
-    return [this.variant, this.size ? 'size-' + this.size : false].filter(Boolean).join(" ")
+    return [this.variant, this.size ? this.size : false].filter(Boolean).join(" ")
   }
 }