+++ libs/extract/src/lib/dropdown/dropdown.ts
@@ -13,7 +13,6 @@ import {
   validate,
   search,
   selectByValue,
-  resetTouchedProperty,
 } from './reducers'
 import { fromEvent, merge, Subject } from 'rxjs'
 import { take, takeUntil } from 'rxjs/operators'
@@ -70,8 +69,6 @@ export const createDropdown = (
   handler.update = (props) => update(handler, listener, create(props))
   handler.validate = (validator: IValidator) =>
     update(handler, listener, validate(handler.dropdown, validator))
-  handler.resetTouchedState = () =>
-    update(handler, listener, resetTouchedProperty(handler.dropdown))
 
   fromEvent(toggler, 'blur')
     .pipe(takeUntil(handler.onDestroy$))