+++ libs/extract/src/lib/id.ts
@@ -1,2 +1,2 @@
 export const randomId = (): string =>
-  'gds-' + Math.random().toString(36).substring(2, 9)
+  'gds--' + Math.random().toString(36).substring(2, 9)