+++ libs/react/src/lib/icons/index.ts
@@ -1,0 +1,4 @@
+export * from './square-exclamation'
+export * from './square-info'
+export * from './check'
+export * from './chevronDown'