+++ libs/react/src/lib/link/link.tsx
@@ -25,6 +25,7 @@ export const Link = ({
       ...props,
     }
     setAnchorProps(newProps)
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [button])
 
   return <a {...anchorProps}>{children}</a>