+++ libs/charts/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.44...@sebgroup/green-charts@0.0.1) (2023-01-27)
+
+### Dependency Updates
+
+* `chlorophyll` updated to version `1.0.0`
+
+
+# [1.0.0-beta.44](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.43...@sebgroup/green-charts@1.0.0-beta.44) (2023-01-26)
+
+
+
+# [1.0.0-beta.43](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.42...@sebgroup/green-charts@1.0.0-beta.43) (2023-01-26)
+
+
+
+# [1.0.0-beta.42](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.41...@sebgroup/green-charts@1.0.0-beta.42) (2023-01-12)
+
+
+
+# [1.0.0-beta.41](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.40...@sebgroup/green-charts@1.0.0-beta.41) (2022-12-23)
+
+
+
 # [1.0.0-beta.40](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.39...@sebgroup/green-charts@1.0.0-beta.40) (2022-12-12)